// src/DockiUp.Frontend/src/app/api/model/projectOriginType.ts
/**
 * DockiUp.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ProjectOriginType = 'Unknown' | 'Git' | 'Compose' | 'Import';

export const ProjectOriginType = {

    Unknown: 'Unknown' as ProjectOriginType,

    Git: 'Git' as ProjectOriginType,

    Compose: 'Compose' as ProjectOriginType,

    Import: 'Import' as ProjectOriginType
};