// src/DockiUp.Frontend/src/app/api/model/projectUpdateMethod.ts
/**
 * DockiUp.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ProjectUpdateMethod = 'Unknown' | 'Webhook' | 'Manual' | 'Periodically';

export const ProjectUpdateMethod = {

    Unknown: 'Unknown' as ProjectUpdateMethod,

    Webhook: 'Webhook' as ProjectUpdateMethod,

    Manual: 'Manual' as ProjectUpdateMethod,

    Periodically: 'Periodically' as ProjectUpdateMethod
};